requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
ultralytics
opencv-python-headless
pandas
python-dotenv
Werkzeug
